healthcare-pro-agents/requirements.txt
google-adk
python-dotenv
requests
fastapi
uvicorn
